birth-chart-react/back-end/requirements.txt
fastapi
uvicorn
pyswisseph
python-dateutil